fastapi/requirements.txt
fastapi
uvicorn
netmiko
requests
python-dotenv
